+++ packages/icons/src/index.ts
@@ -1,6 +1,3 @@
-export { ReactComponent as Alert } from './icons/alert.svg';
-export { default as AlertRaw } from './icons/alert.svg?raw';
-
 export { ReactComponent as AngleDown } from './icons/angle-down.svg';
 export { default as AngleDownRaw } from './icons/angle-down.svg?raw';
 
@@ -52,6 +49,9 @@ export { default as ExclamationCircleRaw } from './icons/exclamation-circle.svg?
 export { ReactComponent as ExclamationTriangle } from './icons/exclamation-triangle.svg';
 export { default as ExclamationTriangleRaw } from './icons/exclamation-triangle.svg?raw';
 
+export { ReactComponent as ExternalLinkAlt } from './icons/external-link-alt.svg';
+export { default as ExternalLinkAltRaw } from './icons/external-link-alt.svg?raw';
+
 export { ReactComponent as EyeSlash } from './icons/eye-slash.svg';
 export { default as EyeSlashRaw } from './icons/eye-slash.svg?raw';
 
@@ -76,6 +76,9 @@ export { default as InfoCircleRaw } from './icons/info-circle.svg?raw';
 export { ReactComponent as KeySkeleton } from './icons/key-skeleton.svg';
 export { default as KeySkeletonRaw } from './icons/key-skeleton.svg?raw';
 
+export { ReactComponent as LinkBroken } from './icons/link-broken.svg';
+export { default as LinkBrokenRaw } from './icons/link-broken.svg?raw';
+
 export { ReactComponent as LockAlt } from './icons/lock-alt.svg';
 export { default as LockAltRaw } from './icons/lock-alt.svg?raw';
 
@@ -90,12 +93,6 @@ export { default as MoonRaw } from './icons/moon.svg?raw';
 
 export { ReactComponent as Newspaper } from './icons/newspaper.svg';
 export { default as NewspaperRaw } from './icons/newspaper.svg?raw';
-
-export { ReactComponent as OpenInNew } from './icons/open-in-new.svg';
-export { default as OpenInNewRaw } from './icons/open-in-new.svg?raw';
-
-export { ReactComponent as Paste } from './icons/paste.svg';
-export { default as PasteRaw } from './icons/paste.svg?raw';
 
 export { ReactComponent as PauseCircle } from './icons/pause-circle.svg';
 export { default as PauseCircleRaw } from './icons/pause-circle.svg?raw';
@@ -136,8 +133,8 @@ export { default as StopCircleRaw } from './icons/stop-circle.svg?raw';
 export { ReactComponent as Sun } from './icons/sun.svg';
 export { default as SunRaw } from './icons/sun.svg?raw';
 
-export { ReactComponent as SyncProblem } from './icons/sync-problem.svg';
-export { default as SyncProblemRaw } from './icons/sync-problem.svg?raw';
+export { ReactComponent as SyncExclamation } from './icons/sync-exclamation.svg';
+export { default as SyncExclamationRaw } from './icons/sync-exclamation.svg?raw';
 
 export { ReactComponent as Times } from './icons/times.svg';
 export { default as TimesRaw } from './icons/times.svg?raw';
@@ -157,6 +154,9 @@ export { default as BitcoinRaw } from './networks/bitcoin.svg?raw';
 export { ReactComponent as Cardano } from './networks/cardano.svg';
 export { default as CardanoRaw } from './networks/cardano.svg?raw';
 
+export { ReactComponent as Doge } from './networks/doge.svg';
+export { default as DogeRaw } from './networks/doge.svg?raw';
+
 export { ReactComponent as Ergo } from './networks/ergo.svg';
 export { default as ErgoRaw } from './networks/ergo.svg?raw';
 
@@ -172,11 +172,11 @@ export { default as LaceRaw } from './wallets/lace.svg?raw';
 export { ReactComponent as MetaMask } from './wallets/meta-mask.svg';
 export { default as MetaMaskRaw } from './wallets/meta-mask.svg?raw';
 
+export { ReactComponent as Mydoge } from './wallets/mydoge.svg';
+export { default as MydogeRaw } from './wallets/mydoge.svg?raw';
+
 export { ReactComponent as Nautilus } from './wallets/nautilus.svg';
 export { default as NautilusRaw } from './wallets/nautilus.svg?raw';
 
 export { ReactComponent as Okx } from './wallets/okx.svg';
 export { default as OkxRaw } from './wallets/okx.svg?raw';
-
-export { ReactComponent as Vespr } from './wallets/vespr.svg';
-export { default as VesprRaw } from './wallets/vespr.svg?raw';