+++ wallets/metamask/src/wallet.ts
@@ -22,7 +22,7 @@ import { BrowserProvider, Contract } from 'ethers';
 
 import { WalletConfig } from './types';
 
-export class MetaMaskWallet implements Wallet {
+export class MetaMaskWallet extends Wallet {
   icon = MetaMaskIcon;
 
   name = 'MetaMask';
@@ -51,6 +51,7 @@ export class MetaMaskWallet implements Wallet {
   }
 
   private get provider() {
+    this.requireAvailable();
     const provider = this.api.getProvider();
 
     if (!provider) throw new InteractionError(this.name);
@@ -58,7 +59,9 @@ export class MetaMaskWallet implements Wallet {
     return provider;
   }
 
-  constructor(private config: WalletConfig) {}
+  constructor(private config: WalletConfig) {
+    super();
+  }
 
   private async permissions() {
     return (await this.provider.request({
@@ -68,6 +71,7 @@ export class MetaMaskWallet implements Wallet {
   }
 
   async connect(): Promise<void> {
+    this.requireAvailable();
     try {
       await this.api.connect();
     } catch (error) {
@@ -76,6 +80,7 @@ export class MetaMaskWallet implements Wallet {
   }
 
   async disconnect(): Promise<void> {
+    this.requireAvailable();
     try {
       await this.api.disconnect();
     } catch (error) {
@@ -100,11 +105,9 @@ export class MetaMaskWallet implements Wallet {
 
     const tokenMap = await this.config.getTokenMap();
 
-    const tokenId = token[tokenMap.getIdKey(this.currentChain)];
-
     let amount;
 
-    if (token.metaData.type === 'native') {
+    if (token.type === 'native') {
       amount = await this.provider.request<string>({
         method: 'eth_getBalance',
         params: [address, 'latest'],
@@ -113,7 +116,7 @@ export class MetaMaskWallet implements Wallet {
       const browserProvider = new BrowserProvider(window.ethereum!);
 
       const contract = new Contract(
-        tokenId,
+        token.tokenId,
         tokenABI,
         await browserProvider.getSigner(),
       );
@@ -124,7 +127,7 @@ export class MetaMaskWallet implements Wallet {
     if (!amount) return 0n;
 
     const wrappedAmount = tokenMap.wrapAmount(
-      token[tokenMap.getIdKey(this.currentChain)],
+      token.tokenId,
       BigInt(amount),
       this.currentChain,
     ).amount;
@@ -183,12 +186,8 @@ export class MetaMaskWallet implements Wallet {
       params.bridgeFee.toString(),
     );
 
-    const tokenMap = await this.config.getTokenMap();
-
-    const tokenId = params.token[tokenMap.getIdKey(this.currentChain)];
-
     const transactionParameters = await this.config.generateTxParameters(
-      tokenId,
+      params.token.tokenId,
       params.lockAddress,
       address,
       params.amount,